move all filters and related components to one file
